pyproject.toml
[project]
name = "greprag"
version = "0.2.0"
description = "AI-powered RAG agent for searching and synthesizing information from PDFs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv==1.2.1",
    "openai==2.15.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "ruff==0.14.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: tests requiring external tools (ugrep, pdftotext)",
]
